+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/jws/JwsContentTypeConstants.kt
@@ -18,7 +18,4 @@ object JwsContentTypeConstants {
     const val CLIENT_ATTESTATION_JWT = "oauth-client-attestation+jwt"
     /** OAuth 2.0 Attestation-Based Client Authentication */
     const val CLIENT_ATTESTATION_POP_JWT = "oauth-client-attestation-pop+jwt"
-    const val DIDCOMM_PLAIN_JSON = "didcomm-plain+json"
-    const val DIDCOMM_SIGNED_JSON = "didcomm-signed+json"
-    const val DIDCOMM_ENCRYPTED_JSON = "didcomm-encrypted+json"
 }